Make the SQL for subscriptions much faster to run.
The PostgreSQL query planner knows to short-circuit some operators such
as IN or ANY such that `foo_id IN (SELECT foo.id FROM foo WHERE ...)`
doesn't need to perform a full table scan on foo.

This change ensures that `user_rel` WHERE clauses use the `IN` operator
to take advantage of the greatly improved performance.